create mailer module to avoid redundant code
