+++ app/mailers/cp_mailer.rb
@@ -1,6 +1,7 @@
 class CpMailer < ApplicationMailer
   default from: ENV['EMAIL_USER']
   layout "mailer"
+  include Mailer
 
   def contract_email(offer, link)
     @offer = offer
@@ -54,23 +55,6 @@ class CpMailer < ApplicationMailer
       return "rd"
     else
       return "th"
-    end
-  end
-
-  def mail_to(email, subject)
-    email = get_email(email)
-    mail(to: email, subject: subject)
-  end
-
-  def get_email(email)
-    if ENV['RAILS_ENV'] == 'development'
-      return ENV['RECIPIENT']
-    elsif ENV['RAILS_ENV'] == 'production'
-      if ENV['domain'].include?('sleepy')
-        return ENV['RECIPIENT']
-      else
-        return email
-      end
     end
   end
